# openapi.yaml
openapi: 3.1.0
info:
  title: Collaborative Travel Planner API
  description: |-
    This is the API for the Collaborative Travel Planner application. It allows users to create, manage, and share travel itineraries. The API supports real-time collaboration features through WebSockets.
    Authentication is handled via JWT Bearer tokens obtained from Google OAuth.
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /trips/{tripId}/agents/optimize-itinerary:
    post:
      summary: Trigger AI agents to optimize trip itinerary
      description: Orchestrates multiple AI agents to analyze requirements and optimize the trip itinerary based on chat messages and preferences.
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    description: Chat messages to analyze
                options:
                  type: object
                  properties:
                    prioritizeTime:
                      type: boolean
                    prioritizeCost:
                      type: boolean
                    prioritizeExperience:
                      type: boolean
                    generateAlternatives:
                      type: boolean
              required:
                - messages
      responses:
        '200':
          description: Itinerary optimization completed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentOptimizationResponse'
        '400':
          description: Invalid request data
        '404':
          description: Trip not found
        '500':
          description: Optimization failed
...
  /trips/{tripId}/agents/analyze-requirements:
    post:
      summary: Analyze travel requirements from chat messages
      description: Uses AI agents to extract and analyze travel requirements from conversation messages.
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messages:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Extracted travel intentions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractedInfo'
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    description: Chat messages to analyze
              required:
                - messages
      responses:
        '200':
          description: Requirements analysis completed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRequirementResponse'
  /trips/{tripId}/agents/adjust-plan:
    post:
      summary: Handle adaptive adjustments for trip disruptions
      description: Creates adjustment plans when disruptions occur during travel.
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                disruption:
                  type: object
                  properties:
                    type:
                      type: string
                      enum: [weather, transportation, accommodation, activity_closure, emergency]
                    severity:
                      type: string
                      enum: [low, medium, high, critical]
                    affectedDestination:
                      type: string
                    affectedDate:
                      type: string
                      format: date
                    description:
                      type: string
                    suggestedActions:
                      type: array
                      items:
                        type: string
                  required:
                    - type
                    - severity
                    - description
              required:
                - disruption
      responses:
        '200':
          description: Adjustment plan created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentAdjustmentResponse'
  /trips/{tripId}/agents/status/{sessionId}:
    get:
      summary: Get agent session status
      description: Retrieves the current status of an AI agent workflow session.
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session status retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSessionStatus'
        '404':
          description: Session not found
  /trips/{tripId}/agents/sessions/{sessionId}:
    delete:
      summary: Cancel an active agent session
      description: Cancels an active AI agent workflow session.
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session cancelled successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  sessionId:
                    type: string
        '400':
          description: Session could not be cancelled
        '404':
          description: Session not found
  /agents/available:
    get:
      summary: Get available AI agents and their capabilities
      description: Returns a list of all available AI agents and their capabilities.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Available agents retrieved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/AgentInfo'
                  totalAgents:
                    type: integer
                  generatedAt:
                    type: string
                    format: date-time
  /agents/sessions:
    get:
      summary: List all active agent sessions
      description: Returns a list of all active AI agent workflow sessions across all trips for the authenticated user.
      tags:
        - AI Agents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Active sessions retrieved successfully.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AgentSessionStatus'
        '500':
          description: Failed to retrieve sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agents/sessions/{sessionId}:
    get:
      summary: Get detailed session information
      description: Retrieves detailed information about a specific AI agent session.
      tags:
        - AI Agents
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentSessionStatus'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Cancel an agent session
      description: Cancels an active AI agent workflow session.
      tags:
        - AI Agents
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session cancelled successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Session cancelled successfully"
                  sessionId:
                    type: string
        '400':
          description: Session could not be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /agents/sessions/{sessionId}/results:
    get:
      summary: Get session results
      description: Retrieves the results from a completed AI agent session.
      tags:
        - AI Agents
      security:
        - bearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session results retrieved successfully.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/AgentOptimizationResponse'
                  - $ref: '#/components/schemas/AgentRequirementResponse'
                  - $ref: '#/components/schemas/AgentAdjustmentResponse'
        '404':
          description: Session not found or no results available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Session still in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ExtractedInfo:
      type: object
      properties:
        destinations:
          type: array
          items:
            type: string
        dates:
          type: array
          items:
            type: string
            format: date
        budget:
          type: number
        interests:
          type: array
          items:
            type: string
        preferences:
          type: object
        mentions:
          type: object
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Service not found"
        message:
          type: string
          description: Detailed error description
          example: "The requested service could not be found"
      required:
        - error
    # AI Agent Schemas
    AgentOptimizationResponse:
      type: object
      properties:
        success:
          type: boolean
        sessionId:
          type: string
        requirements:
          $ref: '#/components/schemas/TravelRequirements'
        itinerary:
          $ref: '#/components/schemas/OptimizedItinerary'
        alternatives:
          type: array
          items:
            $ref: '#/components/schemas/OptimizedItinerary'
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        processingTime:
          type: number
          description: Processing time in milliseconds
        recommendations:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
    AgentRequirementResponse:
      type: object
      properties:
        success:
          type: boolean
        sessionId:
          type: string
        requirements:
          $ref: '#/components/schemas/TravelRequirements'
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
        processingTime:
          type: number
          description: Processing time in milliseconds
        recommendations:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
    AgentAdjustmentResponse:
      type: object
      properties:
        success:
          type: boolean
        sessionId:
          type: string
        disruption:
          $ref: '#/components/schemas/DisruptionEvent'
        originalItinerary:
          $ref: '#/components/schemas/OptimizedItinerary'
        adjustedItinerary:
          $ref: '#/components/schemas/OptimizedItinerary'
        impactAnalysis:
          type: object
          properties:
            affectedDays:
              type: integer
            costImpact:
              type: number
            experienceImpact:
              type: number
        alternatives:
          type: array
          items:
            $ref: '#/components/schemas/OptimizedItinerary'
        recommendations:
          type: array
          items:
            type: string
        generatedAt:
          type: string
          format: date-time
    AgentSessionStatus:
      type: object
      properties:
        sessionId:
          type: string
        tripId:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, completed, failed, cancelled]
        currentStep:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        errors:
          type: array
          items:
            type: string
        hasResults:
          type: boolean
    AgentInfo:
      type: object
      properties:
        name:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            version:
              type: string
            capabilities:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  description:
                    type: string
                  inputSchema:
                    type: object
                  outputSchema:
                    type: object
    TravelRequirements:
      type: object
      properties:
        destinations:
          type: array
          items:
            type: string
        budget:
          type: object
          properties:
            total:
              type: number
            currency:
              type: string
            categories:
              type: object
        dates:
          type: object
          properties:
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
            duration:
              type: integer
            flexibility:
              type: string
              enum: [rigid, flexible, very_flexible]
        preferences:
          type: object
        groupDynamics:
          type: object
          properties:
            size:
              type: integer
            interests:
              type: array
              items:
                type: string
        priorities:
          type: object
          properties:
            primary:
              type: array
              items:
                type: string
            secondary:
              type: array
              items:
                type: string
        sentiment:
          type: object
          properties:
            overall:
              type: string
              enum: [positive, neutral, negative]
            excitement:
              type: number
              format: float
            consensus:
              type: number
              format: float
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
    OptimizedItinerary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        totalCost:
          type: number
        currency:
          type: string
        duration:
          type: integer
        destinations:
          type: array
          items:
            type: object
        dailyPlans:
          type: array
          items:
            type: object
        transportation:
          type: array
          items:
            type: object
        budgetBreakdown:
          type: object
          properties:
            accommodation:
              type: number
            transportation:
              type: number
            activities:
              type: number
            food:
              type: number
            miscellaneous:
              type: number
        optimizationScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
    DisruptionEvent:
      type: object
      properties:
        type:
          type: string
          enum: [weather, transportation, accommodation, activity_closure, emergency]
        severity:
          type: string
          enum: [low, medium, high, critical]
        affectedDestination:
          type: string
        affectedDate:
          type: string
          format: date
        description:
          type: string
        suggestedActions:
          type: array
          items:
            type: string
    # Version Control Schemas
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT